requirements.txt
python-telegram-bot==13.10
aiosqlite==0.19.0
python-dotenv==1.0.0
cryptography==41.0.3
aiohttp==3.7.4
pydantic==1.10.8 